refactor(queries.js): rename the file and extract doctype to constant
