+++ src/components/Hooks/useService.jsx
@@ -5,6 +5,7 @@ import { useClient } from 'cozy-client'
 import log from 'cozy-logger'
 
 import { fetchNormalizedServiceByName } from '../../helpers/fetches'
+import { DOCTYPE_TRIGGERS } from '../../helpers/doctypes'
 
 const hasServiceBeenLaunched = service => {
   return get(service, 'current_state.last_success', '').length > 0
@@ -28,7 +29,7 @@ const useService = name => {
     // start service if it has never been launched before
     if (service && hasBeenLaunched === false) {
       log('info', `Executing ${name} service by Contacts app`)
-      client.collection('io.cozy.triggers').launch(service)
+      client.collection(DOCTYPE_TRIGGERS).launch(service)
     }
   }, [service, client, hasBeenLaunched, name])
