+++ src/components/Modals/ContactFormModal.jsx
@@ -6,7 +6,7 @@ import { FixedDialog } from 'cozy-ui/transpiled/react/CozyDialogs'
 import Alerter from 'cozy-ui/transpiled/react/Alerter'
 import Button from 'cozy-ui/transpiled/react/Buttons'
 import { useI18n } from 'cozy-ui/transpiled/react/I18n'
-import { buildContactQuery } from '../../helpers/queries'
+import { buildContactQuery } from '../../queries/queries'
 
 import SelectedGroupContext from '../Contexts/SelectedGroup'
 import ContactForm, { getSubmitContactForm } from '../ContactCard/ContactForm'
